requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
autoflake==2.3.1
black==24.4.2
certifi==2024.7.4
click==8.1.7
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.2.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomlkit==0.13.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
